use int for mmsi instead of long

+++ ais/src/main/java/au/gov/amsa/ais/HasMmsi.java
@@ -2,5 +2,5 @@ package au.gov.amsa.ais;
 
 public interface HasMmsi {
 
-	long getMmsi();
+    int getMmsi();
 }
