+++ animator/src/main/java/au/gov/amsa/animator/Model.java
@@ -14,7 +14,7 @@ public interface Model {
      * 
      * @return
      */
-    Map<Long, Collection<Fix>> recent();
+    Map<Integer, Collection<Fix>> recent();
 
     long stepNumber();
